fix: print start of partition on next_free check (#85)
If a partition starts too early an error occurs and a message is logged
that should show where the partition starts and where it is allowed to
start. Previously the log message showed the next allowed block for both
values. Change this to show where the partition starts and where it is
allowed to start.

Fixes: c8902e2 ("chore: polish image partition table computation")

Signed-off-by: Thomas Böhler <t.boehler@kunbus.com>